### 2019-07-18 #1581 @whedon
[PRE REVIEW]: datacleanbot: an automated data cleaning tool
**Submitting author:** @Ji-Zhang (<a href="http://orcid.org/0000-0002-4913-518X">Ji Zhang</a>)
**Repository:** <a href="https://github.com/Ji-Zhang/datacleanbot" target ="_blank">https://github.com/Ji-Zhang/datacleanbot</a>
**Version:** v0.4
