### 2019-07-19 #1581 @usethedata
@Ji-Zhang Thanks for your submission.  In reading through it, one thing I notice is that there's no contributor information.  Part of the expectation is under the heading of community guidelines:

    There should be clear guidelines for third-parties wishing to:

    * Contribute to the software
    * Report issues or problems with the software
    * Seek support

I'm not seeing information about contributions (a very simple contributing.md file can be done, explaining how you want to receive code or documentation contributions.  Reporting issues and seeking support might go well in your documentation.  It doesn't need to be long or elaborate.

Finally, do you have any suggested reviewers?  Thanks
