### 2019-07-19 652c967
Create contributing.md
