+++ contributing.md
@@ -1,0 +1,13 @@
+# How to contribute to ``datacleanbot``
+
+First and foremost: by contributing to this project, you agree to abide by the [Contributor Covenant code of conduct](https://www.contributor-covenant.org/).
+
+Contributions to ``datacleanbot`` are welcomed. Please feel free to create a new issue [here](https://github.com/Ji-Zhang/datacleanbot/issues) in any of the following 4 cases:
+    
+    * Bug Reports
+    * Features
+    * Performance / Optimization
+    * Questions / Seek Support
+
+Please state in the comment which case it falls into.
+
