### 2019-07-19 4c06d5c
Update contributing.md
