### 2019-07-19 #1581 @Ji-Zhang
@usethedata Thank you for reminding that. I just added the contributing.md file.

Regarding the reviewers, I am wondering if @rgmyr or @kellieotto is willing to be the reviewer. The tool is written in Python,  any reviewer familiar with machine learning, data science or data visualization is welcomed.
