### 2019-07-30 #1581 @usethedata
@kellieotto Thanks for agreeing to review.  In the interests of keeping this moving forward, I'm going to go ahead and open up the review issue.  We do need to find a second reviewer, and I'll add that person to the review issue.
