### 2019-07-30 #1608 @whedon
[REVIEW]: datacleanbot: an automated data cleaning tool
**Submitting author:** @Ji-Zhang (<a href="http://orcid.org/0000-0002-4913-518X">Ji Zhang</a>)
**Repository:** <a href="https://github.com/Ji-Zhang/datacleanbot" target ="_blank">https://github.com/Ji-Zhang/datacleanbot</a>
**Version:** v0.4

@kellieotto and @jjmcnelis , please carry out your review in this issue by updating the checklist below. If you cannot edit the checklist please:
