### 2019-08-04 #1608 @usethedata
@kellieotto Just checking in (I've been off-line working a major proposal and getting my house on the market to sell).  Any questions?
