### 2019-08-05 #1608 @kellieotto
Thanks for checking in. I plan to get to it this week, no questions!

On Sun, Aug 4, 2019 at 3:40 PM Bruce Wilson <notifications@github.com>
wrote:

> @kellieotto <https://github.com/kellieotto> Just checking in (I've been
> off-line working a major proposal and getting my house on the market to
> sell). Any questions?
>
