### 2019-08-08 Ji-Zhang/datacleanbot#6 @kellieotto
Paper references
This issue is part of your [JOSS review](https://github.com/openjournals/joss-reviews/issues/1608).

The second reference in your paper bibliography is formatted in a strange way (check the compiled PDF). Please add DOIs for the 2nd-4th references (if they have a DOI).
