### 2019-08-08 Ji-Zhang/datacleanbot#7 @kellieotto
Can't run examples
This issue is part of your [JOSS review](https://github.com/openjournals/joss-reviews/issues/1608).

I can't get the examples (in the README, and in the readthedocs example) to run. First, I had to pip install openml (as it is not installed via the requirements file).

Then I get the following error:
`TypeError: get_data() got an unexpected keyword argument 'return_attribute_names'`
