### 2019-08-08 Ji-Zhang/datacleanbot#8 @kellieotto
Missing unit tests
This issue is part of your [JOSS review](https://github.com/openjournals/joss-reviews/issues/1608).

For JOSS, you are required to have automated testing. I'd like to see [unit tests](https://jeffknupp.com/blog/2013/12/09/improve-your-python-understanding-unit-testing//) for each function. Once those are in place, I recommend setting up [Travis CI](https://travis-ci.org).
