### 2019-08-08 Ji-Zhang/datacleanbot#7 @Ji-Zhang
Thanks! `openml` has been updated. I just updated the README and the readthedocs example. It should work now.
