### 2019-08-08 f860715
update paper reference
