+++ paper/paper.bib
@@ -1,19 +1,24 @@
-@article{vanschoren2014openml,
-  title={OpenML: networked science in machine learning},
-  author={Vanschoren, Joaquin and Van Rijn, Jan N and Bischl, Bernd and Torgo, Luis},
-  journal={ACM SIGKDD Explorations Newsletter},
-  volume={15},
-  number={2},
-  pages={49--60},
-  year={2014},
-  publisher={ACM}
+@article{OpenML2013,
+  author = {Vanschoren, Joaquin and van Rijn, Jan N. and Bischl, Bernd and Torgo, Luis},
+  title = {OpenML: Networked Science in Machine Learning},
+  journal = {SIGKDD Explorations},
+  volume = {15},
+  number = {2},
+  year = {2013},
+  pages = {49--60},
+  url = {http://doi.acm.org/10.1145/2641190.2641198},
+  doi = {10.1145/2641190.2641198},
+  publisher = {ACM},
+  address = {New York, NY, USA},
 }
 
-@misc{Sessions_theeffects,
-    author = {Valerie Sessions and Marco Valtorta},
-    title = {THE EFFECTS OF DATA QUALITY ON MACHINE LEARNING ALGORITHMS (Research-in-Progress – IQ Concepts, Tools, Metrics, Measures, Models, and Methodologies)},
-    year = {}
+@inproceedings{Sessions2006TheEO,
+  title={The Effects of Data Quality on Machine Learning Algorithms},
+  author={Valerie Sessions and Marco Valtorta},
+  booktitle={ICIQ},
+  year={2006}
 }
+
 
 @article{vergari2018automatic,
   title={Automatic Bayesian Density Analysis},
@@ -22,10 +27,21 @@
   year={2018}
 }
 
-@misc{bilogur2018missingno,
-    author = "{Aleksey Bilogur}",
-    title = "Missingno: a missing data visualization suite",
-    year = "2018",
-    url = "https://doi.org/10.21105/joss.00547",
-    journal={Open Source Software}
- }
+
+@article{Bilogur2018Missingno,
+  journal = {The Journal of Open Source Software},
+  doi = {10.21105/joss.00547},
+  issn = {2475-9066},
+  number = {22},
+  publisher = {The Open Journal},
+  title = {Missingno: a missing data visualization suite},
+  url = {http://dx.doi.org/10.21105/joss.00547},
+  volume = {3},
+  author = {Bilogur, Aleksey},
+  pages = {547},
+  date = {2018-02-27},
+  year = {2018},
+  month = {2},
+  day = {27},
+}
+
