### 2019-08-09 a4b4eef
Update usage.rst
