### 2019-08-13 Ji-Zhang/datacleanbot#8 @Ji-Zhang
Sorry I was not aware of the unit test requirement. My intention was to test the package by these [two examples](https://datacleanbot.readthedocs.io/en/latest/example.html) and they work nicely so far. I will implement the tests progressively but it may take some time.
