### 2019-08-13 Ji-Zhang/datacleanbot#8 @kellieotto
Sounds good. Please drop me a line when tests are in place so I can take a look.
