### 2019-09-18 #1608 @kthyng
Hi @Ji-Zhang how are things going on your end for your JOSS submission?
