### 2019-09-25 #1608 @Ji-Zhang
Hi @jjmcnelis I apologize for the delay to answer your questions.

The following are my answers:

1. Getting started in fresh Python 3 environments:
- rpy2 is related to the data type discovery functionality, which is based on the [Bayesian model](https://github.com/Ji-Zhang/datacleanbot/tree/master/datacleanbot/bayesian). Without rpy2, the Bayesian model can not run successfully.
- Only openml is recommended to be installed manually. The others are not deliberate. Could you please let me know which submodels you need to install manually? Thanks!

2. The remote file metalearner is a pre-trained machine learning model. The prediction of the optimal algorithm involves this model. However, the pickle file metalearner cannot be packaged together through pip. So I took this way around.
The sklearn incompatible problem is related to GridSearchCV which I used to train the metalearner model. GridSearch had a DeprecationDict in it, but that doesn't exist in sklearn any more. I tried to retrain the model with the newest sklearn version but it doesn't seem work.

3. Some dependencies of datacleanbot only support Python3 so I assume Python2 won't work. I will add this note to the document.

Please feel free to let me know if you have more questions. Again, sorry for the delay.
