### 2019-10-22 Ji-Zhang/datacleanbot#8 @kthyng
Hi @kellieotto and @Ji-Zhang — here is what the author guidelines say: https://joss.readthedocs.io/en/latest/review_criteria.html#tests

Authors are encouraged but not required to have testing such as via continuous integration. Some sort of objective testing is required.
