### 2019-10-22 #1608 @kthyng
@jjmcnelis How has @Ji-Zhang responded to your concerns?
