### 2019-10-22 Ji-Zhang/datacleanbot#7 @kellieotto
@Ji-Zhang @kthyng this fixed the original problem, but I am still not able to run examples. I get the following error (and have tried a few different openml datasets):

```
LoweringError: Failed in object mode pipeline (step: object mode frontend)
Failed in object mode pipeline (step: object mode backend)
(<class 'numba.ir.StaticSetItem'>, X_hat[(slice(None, None, None), slice(None, None, None))] = $0.13)

File "../../../../../../opt/conda/lib/python3.6/site-packages/datacleanbot/bayesian/explain.py", line 355:
def predict_data(X, C, R, S,
    <source elided>
    X_hat = np.zeros((N, D))
    X_hat[:, :] = np.inf
    ^

[1] During: lowering "X_hat[(slice(None, None, None), slice(None, None, None))] = $0.13" at /opt/conda/lib/python3.6/site-packages/datacleanbot/bayesian/explain.py (355)
```
