### 2019-10-22 Ji-Zhang/datacleanbot#8 @kellieotto
@kthyng Right now, I cannot find code that is directly testing the package. As @Ji-Zhang pointed out there are some examples, but when I try to run them they suffer from the error I pointed out in #7
