### 2019-10-22 Ji-Zhang/datacleanbot#8 @kthyng
@kellieotto Ok makes sense. If the examples are not adequate to objectively test the code, once they work, then it would make sense to pursue this issue.
