### 2019-10-24 Ji-Zhang/datacleanbot#7 @Ji-Zhang
Hi @kellieotto, may I ask what version of the library `numba` you are using? Thanks.
