### 2019-11-03 #1608 @jjmcnelis
@kthyng @Ji-Zhang I'm sorry for holding you up. Thanks for the thoughtful responses to my questions. I'm satisfied with points 1 and 2. Point 3 needs addressing.

1. I raised the point about `rpy2` in the interest of usability. I don't use R often enough to justify having it installed on my workstation. I'd look into implementing the functionality provided by `rpy2` in Python in a future release so that your software will be usable by the widest audience. But this isn't a JOSS requirement as far as I know.

2. Since my initial review I've come across a few other examples of machine learning resources retrieved remotely when a model runs, so I guess it's reasonably common.

3. Dependency issues are many, on my end. I haven't been able to run the examples in three separate attempts, on three separate machines. And I get a number of different tracebacks that bear no resemblance to each other. Did you hear from @kellieotto about their dependency issues (#issuecomment-519314651)?
