### 2019-11-24 #1608 @usethedata
@Ji-Zhang what are your thoughts about the question in point 3 above?
