### 2019-11-27 #1608 @Ji-Zhang
@usethedata I am working on simplifying the package such that the dependencies can become not this heavy. The light version should have been released before the weekend.
