### 2019-11-28 ae83f6c
remove Bayesian part
