### 2019-11-28 20630ab
remove Bayesian data type discover
