+++ datacleanbot/dataclean.py
@@ -19,7 +19,7 @@ from sklearn.impute import SimpleImputer
 from sklearn.pipeline import Pipeline
 from sklearn.ensemble import RandomForestClassifier
 from scipy.io import savemat
-from datacleanbot.bayesian.bin import abda
+# from datacleanbot.bayesian.bin import abda
 
 from sklearn.preprocessing import Imputer
 from fancyimpute import KNN, IterativeImputer, MatrixFactorization
@@ -288,83 +288,84 @@ def discover_type_heuristic(data):
         #print ("Result:",inferredType) #For debugging purposes
     return result
 
-def generate_mat(Xy, extra_cardinality=1):
-    """Convert data to mat format.
+# def generate_mat(Xy, extra_cardinality=1):
+#     """Convert data to mat format.
 
-    In order to use the Bayesian model, data need to be converted
-    to the .mat format.
-    """
-    data = Xy
-    simple_types = discover_type_heuristic(data)
-    # map simple types to meta types
-    # 1: real (w positive: all real | positive | interval)
-    # 2: real (w/o positive: all real | interval)
-    # 3: binary data
-    # 4: discrete (non-binary: categorical | ordinal | count)
-    # note: for now, the implemented bayesian method version by isabel can distinguish between real, postive real, categorical and count
-    # the binary data should be mapped to meta type 4 discrete instead of meta type 3 due to the limited implemented version. This may change
-    # if the extended version has been implemented by isabel.
-    meta_types = [] 
-    for i in range(len(simple_types)):
-#         print(simple_types[i])
-        if simple_types[i] == "bool":
-            meta_types.append(4) # may change in the future
-        elif simple_types[i] == "int64" or simple_types[i] == "float64":
-            if (len(np.unique(data[:,i])) < 0.02 * len(data[:,i]) and \
-                len(np.unique(data[:,i])) < 50):
-                meta_types.append(4)
-            else:
-                if (data[:,i] > 0).all():
-                    meta_types.append(1)
-                else:
-                    meta_types.append(2)
-        else:
-            meta_types.append(1)
-    discrete_cardinality = [] # max for discrete feature, 1 for others
-    for i in range(len(meta_types)):
-        if (meta_types[i] == 4):
-            discrete_cardinality.append(int(np.max(data[:,i])) + extra_cardinality) 
-        else:
-            discrete_cardinality.append(1)
-    data_dict = {'X': data,
-                 'T': np.asarray(meta_types),
-                 'R': np.asarray(discrete_cardinality)}
-    # pprint.pprint(data_dict)
-    savemat('data.mat', data_dict, oned_as='row')
+#     In order to use the Bayesian model, data need to be converted
+#     to the .mat format.
+#     """
+#     data = Xy
+#     simple_types = discover_type_heuristic(data)
+#     # map simple types to meta types
+#     # 1: real (w positive: all real | positive | interval)
+#     # 2: real (w/o positive: all real | interval)
+#     # 3: binary data
+#     # 4: discrete (non-binary: categorical | ordinal | count)
+#     # note: for now, the implemented bayesian method version by isabel can distinguish between real, postive real, categorical and count
+#     # the binary data should be mapped to meta type 4 discrete instead of meta type 3 due to the limited implemented version. This may change
+#     # if the extended version has been implemented by isabel.
+#     meta_types = [] 
+#     for i in range(len(simple_types)):
+# #         print(simple_types[i])
+#         if simple_types[i] == "bool":
+#             meta_types.append(4) # may change in the future
[…]
