+++ setup.py
@@ -25,6 +25,6 @@ setuptools.setup(
         'fancyimpute'],
     classifiers=[
         "Programming Language :: Python :: 3",
-        "License :: OSI Approved :: MIT License"
+        "License :: OSI Approved :: MIT License",
     ],
 )
