+++ setup.py
@@ -15,9 +15,9 @@ setuptools.setup(
     packages=setuptools.find_packages(),
     install_requires=[ 
         'numpy>=1.14.2',
-        'pandas',
-        'scikit-learn>=0.20.0', 
-        'scipy>=1.0.0',
+        'pandas>=0.19.0, <=0.23.1',
+        'scikit-learn>=0.20.0, <0.21.0', 
+        'scipy>=1.2.1',
         'seaborn>=0.8',
         'matplotlib>=2.2.2',
         'missingno>=0.4.0',
