### 2019-11-29 8ffbab3
update doc after deleting bayesian
