+++ paper/paper.md
@@ -30,7 +30,7 @@ Data in real life almost never come in a clean way, and poor data quality may se
     
     * The most important features
     * Statistical information (e.g., mean, max, min)
-    * **Data types of features**
+    * Data types of features
 
 * Clean common data problems in the raw dataset
     
@@ -39,9 +39,7 @@ Data in real life almost never come in a clean way, and poor data quality may se
     * **Missing values**
     * **Outliers**
 
-The three aspects ``datacleanbot`` meaningfully automates are marked in bold.
-
-**Data types of features**: ``datacleanbot`` detects both basic data types (bool, date, float, integer and string) and statistical data types (real-valued, positive real-valued, count and categorical). ``datacleanbot`` detects basic data types by applying some logical rules. For statistical data types, ``datacleanbot`` utlizes the Bayesian model abda [@vergari2018automatic].
+The two aspects ``datacleanbot`` meaningfully automates are marked in bold.
 
 **Missing values**: ``datacleanbot`` identifies characters 'n/a', 'na', '--', '?' as missing. Users can add extra characters to be identified as missing. After the missing values being detected, ``datacleanbot`` presents the missing data in effective visualizations with the help of missingno [@Bilogur2018Missingno]. Afterwards, ``datacleanbot`` recommends the appropriate approach to delete or impute missing values according to the missing mechanism of the given dataset.
 
