+++ paper/paper.bib
@@ -19,15 +19,6 @@
   year={2006}
 }
 
-
-@article{vergari2018automatic,
-  title={Automatic Bayesian Density Analysis},
-  author={Vergari, Antonio and Molina, Alejandro and Peharz, Robert and Ghahramani, Zoubin and Kersting, Kristian and Valera, Isabel},
-  journal={arXiv preprint arXiv:1807.09306},
-  year={2018}
-}
-
-
 @article{Bilogur2018Missingno,
   journal = {The Journal of Open Source Software},
   doi = {10.21105/joss.00547},
