### 2019-11-29 Ji-Zhang/datacleanbot#7 @Ji-Zhang
I have deprecated the feature of detecting feature data type using the Bayesian method as too many dependencies there. I created a new clean virtual environment to test it and it works well. I hope it can work on your end now. Thank you.
