### 2020-02-01 #1608 @arfon
:wave: @kellieotto, @jjmcnelis, in [their most recent comment](https://github.com/openjournals/joss-reviews/issues/1608#issuecomment-559885510) it sounds like @Ji-Zhang has simplified some of the requirements here.

Do you think you could both take another look at this software again and confirm you're able to install it now? If at all possible, I'd love to get this review wrapped up in the next couple of weeks.
