### 2020-02-05 Ji-Zhang/datacleanbot#10 @kellieotto
Cannot import package due to sklearn dependency
This issue is part of your [JOSS review](https://github.com/openjournals/joss-reviews/issues/1608).

I was able to install the package but can't import it. I get the following error

```
ImportError: cannot import name 'Imputer' from 'sklearn.preprocessing' (/opt/conda/envs/py3-primary/lib/python3.7/site-packages/sklearn/preprocessing/__init__.py)
```

seems related to this? https://stackoverflow.com/questions/59439096/importerror-cannnot-import-name-imputer-from-sklearn-preprocessing
