### 2020-02-16 Ji-Zhang/datacleanbot#10 @Ji-Zhang
Hi @kellieotto , this bug has been fixed. Please let me know if other problems. Thanks.
