### 2020-02-16 #1608 @Ji-Zhang
Hi @arfon and @kellieotto , thank you for spotting the bug. This bug is now fixed. Please update to the latest version of `datacleanbot` and feel free to let me know if any questions. Thank you in advance.
