### 2020-03-08 Ji-Zhang/datacleanbot#10 @Ji-Zhang
Hi @kellieotto, may I ask in which environment you are using the TensorFlow? There is not much I can do in the package to fix this.

If you are using TensorFlow with GPU, you need to install CUDA and cuDNN. Please follow instructions on https://www.tensorflow.org/install/

If you have already install CUDA and cuDNN, but still get this error, then you probably forgot to export your libraries: for Linux, you may need to set LD_LIBRARY_PATH to include CUDA libraries.

If the above can not fix the problem, please let me know so I can further help. Thanks.
