### 2020-05-10 Ji-Zhang/datacleanbot#10 @kellieotto
Hi @Ji-Zhang, sorry for the huge time delay between responses on this.

I have installed everything now. I am running the code example you have in the README and get this error:
```
<ipython-input-15-b513b18ad249> in <module>
      1 import datacleanbot.dataclean as dc
----> 2 Xy = dc.autoclean(Xy, data.name, features)

~/miniconda3/lib/python3.7/site-packages/datacleanbot/dataclean.py in autoclean(Xy, dataset_name, features)
   1345     features = unify_name_consistency(features)
   1346     features_new, Xy_filled = handle_missing(features, Xy)
-> 1347     Xy_cleaned = handle_outlier(features_new, Xy_filled)
   1348     return Xy_cleaned

~/miniconda3/lib/python3.7/site-packages/datacleanbot/dataclean.py in handle_outlier(features, Xy)
   1286     X = Xy[:,:-1]
   1287     y = Xy[:,-1]
-> 1288     best = predict_best_anomaly_algorithm(X, y)
   1289     df = pd.DataFrame(Xy)
   1290     display(HTML('<h4>Visualize Outliers ... </h4>'))

~/miniconda3/lib/python3.7/site-packages/datacleanbot/dataclean.py in predict_best_anomaly_algorithm(X, y)
   1050
   1051     # load meta learner
-> 1052     metalearner = joblib.load(urlopen("https://github.com/Ji-Zhang/datacleanbot/blob/master/process/AutomaticOutlierDetection/metalearner_rf.pkl?raw=true"))
   1053     best_anomaly_algorithm = metalearner.predict(mf)
   1054     if best_anomaly_algorithm[0] == 0:

~/miniconda3/lib/python3.7/site-packages/sklearn/externals/joblib/numpy_pickle.py in load(filename, mmap_mode)
    586         filename = getattr(fobj, 'name', '')
    587         with _read_fileobject(fobj, filename, mmap_mode) as fobj:
--> 588             obj = _unpickle(fobj)
    589     else:
    590         with open(filename, 'rb') as f:

~/miniconda3/lib/python3.7/site-packages/sklearn/externals/joblib/numpy_pickle.py in _unpickle(fobj, filename, mmap_mode)
    524     obj = None
    525     try:
--> 526         obj = unpickler.load()
    527         if unpickler.compat_mode:
    528             warnings.warn("The file '%s' has been generated with a "

~/miniconda3/lib/python3.7/pickle.py in load(self)
   1083                     raise EOFError
   1084                 assert isinstance(key, bytes_types)
-> 1085                 dispatch[key[0]](self)
   1086         except _Stop as stopinst:
   1087             return stopinst.value

~/miniconda3/lib/python3.7/pickle.py in load_global(self)
   1371         module = self.readline()[:-1].decode("utf-8")
   1372         name = self.readline()[:-1].decode("utf-8")
-> 1373         klass = self.find_class(module, name)
   1374         self.append(klass)
   1375     dispatch[GLOBAL[0]] = load_global

~/miniconda3/lib/python3.7/pickle.py in find_class(self, module, name)
   1421             elif module in _compat_pickle.IMPORT_MAPPING:
   1422                 module = _compat_pickle.IMPORT_MAPPING[module]
-> 1423         __import__(module, level=0)
   1424         if self.proto >= 4:
   1425             return _getattribute(sys.modules[module], name)[0]

ModuleNotFoundError: No module named 'sklearn.ensemble._forest'
```
I think it's related to package versions + pickling. [I found this issue that seems related.](https://github.com/scikit-learn/scikit-learn/issues/16570)
