### 2020-06-07 Ji-Zhang/datacleanbot#10 @Ji-Zhang
Hi @kellieotto , this bug should be fixed now. Could you please test it again? Thanks in advance.
