### 2020-06-28 Ji-Zhang/datacleanbot#10 @Ji-Zhang
Hi @kellieotto ， sorry for the inconvenience. I changed the way to load the trained model. Could you please try it again?  `pip install datacleanbot==0.9`
