+++ setup.py
@@ -5,7 +5,7 @@ with open("README.md", "r") as fh:
 
 setuptools.setup(
     name="datacleanbot",
-    version="0.9",
+    version="0.91",
     author="Ji Zhang",
     author_email="",
     description="automated data cleaning tool",
@@ -16,13 +16,13 @@ setuptools.setup(
     install_requires=[ 
         'numpy>=1.14.2',
         'pandas>=0.19.0, <=0.23.1',
-        'scikit-learn>=0.20.0', 
+        'scikit-learn>=0.23.2', 
         'scipy>=1.2.1',
         'seaborn>=0.8',
         'matplotlib>=2.2.2',
         'missingno==0.4.1',
         'openml>=0.9.0',
-        'fancyimpute==0.4.0'],
+        'fancyimpute==0.5.5'],
     classifiers=[
         "Programming Language :: Python :: 3",
         "License :: OSI Approved :: MIT License",
