### 2020-08-30 #1608 @Ji-Zhang
Hi @kellieotto , I think it is because some dependent packages are outdated. I fixed it by updating the setup file. Run `pip install datacleanbot` in a newly created environment, the example should work. Thanks for all the time you have devoted so far.
