### 2020-08-31 #1608 @arfon
Hi @Ji-Zhang. This review has been going on for a very long time (more than 1 year) and both reviewers have struggled with this software. As a result, I've just taken a deeper look at this submission and I have to say that I think this software is still well below the standard that is allowable in JOSS. For example:

- Multiple blocks of commented out code: [here](https://github.com/Ji-Zhang/datacleanbot/blob/a0a6a9a072ab6a96d1b7c43a25d51680c0683c1b/datacleanbot/dataclean.py#L291-L367) • [here](https://github.com/Ji-Zhang/datacleanbot/blob/a0a6a9a072ab6a96d1b7c43a25d51680c0683c1b/datacleanbot/dataclean.py#L391-L396) • [here](https://github.com/Ji-Zhang/datacleanbot/blob/a0a6a9a072ab6a96d1b7c43a25d51680c0683c1b/datacleanbot/dataclean.py#L599-L631) • [here](https://github.com/Ji-Zhang/datacleanbot/blob/a0a6a9a072ab6a96d1b7c43a25d51680c0683c1b/datacleanbot/dataclean.py#L1156-L1168)
- No tests.
- Multiple [unresolved issues](https://github.com/Ji-Zhang/datacleanbot/issues?q=is%3Aopen+is%3Aissue)

Given the age of this submission, and the state of this software currently I'm going to reject this submission. If you are able to substantially improve the quality of this software then we would welcome a resubmission to JOSS at some point in the future.

@kellieotto, @jjmcnelis - thanks for your time here. I'm sorry your work here didn't result in a paper for the author.
